# src/EasyIotSharp.API/StaticFiles/swagger/v1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "service.easyiotsharp：项目服务",
    "version": "v1"
  },
  "paths": {
...
    "/TenantAccount/Soldier/Validate": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "用户登录验证",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateSoldierInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateSoldierInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateSoldierInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateSoldierInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateSoldierOutputUPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateSoldierOutputUPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateSoldierOutputUPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Soldier/Query": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通过条件分页查询用户列表",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/QuerySoldierInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuerySoldierInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuerySoldierInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuerySoldierInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SoldierDtoPagedResultDtoUPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoldierDtoPagedResultDtoUPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoldierDtoPagedResultDtoUPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Soldier/Insert": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "添加一条用户信息",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/InsertSoldierInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertSoldierInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertSoldierInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InsertSoldierInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Soldier/Update": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通过id修改一条用户信息",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Soldier/UpdateIsEnable": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通过id修改一条用户信息的启用状态",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierIsEnableInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierIsEnableInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierIsEnableInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierIsEnableInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Soldier/Delete": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通过id删除一条用户信息",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteSoldierInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteSoldierInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteSoldierInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteSoldierInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Soldier/Role/Update": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通过id修改一条用户的角色信息",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierRoleInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierRoleInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierRoleInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoldierRoleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Menu/Get": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通过id获取一条菜单信息",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string",
              "description": "",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MenuDtoUPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuDtoUPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuDtoUPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Menu/Query": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通通过条件分页查询菜单列表",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/QueryMenuInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryMenuInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryMenuInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QueryMenuInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MenuDtoPagedResultDtoUPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuDtoPagedResultDtoUPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuDtoPagedResultDtoUPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/TenantAccount/Menu/Insert": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "添加一条菜单信息",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/InsertMenuInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertMenuInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertMenuInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InsertMenuInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Menu/Update": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通过id修改一条菜单信息",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Menu/UpdateIsEnable": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通过od修改一条菜单是否启用状态",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIsEnableMenuInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIsEnableMenuInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIsEnableMenuInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIsEnableMenuInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Menu/Delete": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通过id获取一条菜单信息",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMenuInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMenuInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMenuInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMenuInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Role/Menu/Get": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通过id获取一条角色菜单信息",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string",
              "description": "",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetRoleMenuOutputUPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRoleMenuOutputUPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRoleMenuOutputUPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/TenantAccount/Role/Query": {
      "post": {
        "tags": [
          "TenantAccount"
        ],
        "summary": "通通过条件分页查询菜单列表",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRoleInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRoleInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRoleInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRoleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDtoPagedResultDtoUPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDtoPagedResultDtoUPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDtoPagedResultDtoUPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UPrimeResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "fullMessage": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ValidateSoldierInput": {
        "type": "object",
        "properties": {
          "mobile": {
            "type": "string",
            "description": "手机号",
            "nullable": true
          },
          "password": {
            "type": "string",
            "description": "密码",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "用户登录验证的入参类"
      },
...
      "ValidateSoldierOutputUPrimeResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ValidateSoldierOutput"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "fullMessage": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "QuerySoldierInput": {
        "type": "object",
        "properties": {
          "keyword": {
            "type": "string",
            "description": "手机号/用户名",
            "nullable": true
          },
          "isEnable": {
            "type": "integer",
            "description": "是否启用\r\n-1=不参与查询\r\n1=启用\r\n2=禁用",
            "format": "int32"
          },
          "pageIndex": {
            "type": "integer",
            "description": "当前开始页",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "页大小",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "通过条件分页查询用户列表的入参类"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Value Bearer {token}",
        "name": "u-token",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}